add access gates to manage routes

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StorePostRequest;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -47,8 +48,13 @@ class PostController extends Controller
     // edit post
     public function edit($id)
     {
-        $post = Post::findOrFail($id);
-        return view('posts/edit', compact('post'));
+        if(Gate::allows('is_super_admin')){
+            $post = Post::findOrFail($id);
+            return view('posts/edit', compact('post'));
+        }
+        abort(401, 'Unauthorized');
+        // $post = Post::findOrFail($id);
+        // return view('posts/edit', compact('post'));
     }
 
     // update post
@@ -79,7 +85,12 @@ class PostController extends Controller
 
     public function create()
     {
-        return view('posts/create');
+        if(Gate::allows('is_admin')){
+            return view('posts/create');
+        }
+        abort(401, 'Unauthorized');
+        
+        // return view('posts/create');
     }
     
     //store post
@@ -112,10 +123,18 @@ class PostController extends Controller
     //delete post
     public function destroy($id)
     {
-        $post = Post::findOrFail($id);
-        $post->delete();
+        if(Gate::allows('is_admin')) {
+             $post = Post::findOrFail($id);
+             $post->delete();
 
-        return redirect('/posts');
+             return redirect('/posts');
+        }
+        abort(401, 'Unauthorized');
+        //
+        // $post = Post::findOrFail($id);
+        // $post->delete();
+
+        // return redirect('/posts');
     }
 
     // restore post
